refactor: reconsile duplicated text fixtures to conftest
